Remove Prelude.printLn
`Debug.print` already takes a string that all environments append a
newline to. It is a misnomer to call this `print`, because when your
canister runs in the magic cloud, there is no `stdout` to print to.
Maybe they will be collected in some data base at some point where you
can query them.

This should have been `Debug.log`, in analogy to Unix’s `syslog` or the
browser’s `console.log`.

Anyways, `Prelude.printLn` is not used anywhere.

Thanks @sasa-tomic for noticing in
https://github.com/dfinity-lab/dfinity/pull/3936#issuecomment-645219168