Expose new bit operations on NatN and IntN
this goes along with https://github.com/dfinity/motoko/pull/2326
(and should not be merged before). It adds new bit operations, wrapping
arithmetic and conversions to IntN and NatN.

It does not port the rest of base to avoid Word. See #214 for that (and
we will likely pull that in selectively).