Expose new bit operations on NatN and IntN (#217)
this goes along with https://github.com/dfinity/motoko/pull/2326
(and should not be merged before). It adds new bit operations, wrapping
arithmetic and conversions to IntN and NatN.

It does not port the rest of base to avoid Word. See #214 for that (and
we will likely pull that in selectively).

Adds Char.fromNat32 so that users can migrate away from `Char.fromWord32` already.

Uses wrapping operators on Word types. This avoids deprecation warnings in the upcoming
compiler version.