Deprecate Char.to/fromWord32
because of the upcoming deprecation of word types, we need to readicate
`Word` from all other modules’ interfaces.

This uses deprecation annotations to warn about the methods in `Char`
thta we will have to remove when we start warning about the `Word` type
itself.

This can be merged into `master` already, as support for the
`@deprecated` annotation is backward-compatible.