feat: more array utility functions (#564)
This change introduces new utility functions to deal with arrays:

  * Array.slice
  * Array.indexOf, Array.lastIndexOf
  * Array.nextIndexOf, Array.prevIndexOf
  * Text.toArray, Text.toVarArray

Controversial design choices:

* `Array.slice` accepts a half-closed range `[from, to)`. The goal is to
eliminate the need to subtract one from the array size and make
off-by-one errors less likely. Both `Array.slice(a, 0, a.size())` and
`Array.slice(a, 0, 0)` work as expected. I found this approach much more
convenient in practice than inclusive ranges that `Iter.range(from, to)`
generates.

* `Array.nextIndexOf` and `Array.prevIndexOf` are assymetrical. The
starting position in `nextIndexOf` is inclusive, while the starting
position in `prevIndexOf` is exclusive. The goal is again to eliminate
the need to subtract one from the array size when scanning backwards.

---------

Co-authored-by: Claudio Russo <claudio@dfinity.org>